pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Antesser <ahriman.kun@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django-bootstrap4 = "^22.3"
dj-database-url = "^1.3.0"
django = "^4.1.7"
flake8 = "^6.0.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.4.0"
django-filter = "^23.1"
rollbar = "^0.16.3"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
